feat improvement on save delete image

+++ lib/presentation/pages/emoji_pack/widgets/sidebar.dart
@@ -6,7 +6,6 @@ import 'package:dc_universal_emot/presentation/bloc/emoji_pack_bloc.dart';
 import 'package:dc_universal_emot/presentation/pages/emoji_pack/emoji_pack_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:provider/provider.dart';
 
 import 'add_emoji_pack_dialog.dart';
 
@@ -32,26 +31,31 @@ class Sidebar extends StatelessWidget {
                 separatorBuilder: (_, __) => const SizedBox(height: 8),
                 itemBuilder: (_, index) {
                   final emojiPack = state.emojiPacks[index];
+                  final isHovered = ValueNotifier(false);
                   return InkWell(
                     onTap: () => context.read<EmojiPackProvider>().setSelectedEmojiPackIndex(index),
-                    child: Consumer<EmojiPackProvider>(
-                      child: Image.file(
-                        File(emojiPack.emojiPath),
-                      ),
-                      builder: (_, state, child) {
-                        return AnimatedContainer(
-                          duration: const Duration(milliseconds: 200),
-                          decoration: BoxDecoration(
-                            color: darkGray100,
-                            borderRadius: BorderRadius.circular(
-                              state.selectedEmojiPackIndex == index ? 8 : 32,
-                            ),
+                    child: MouseRegion(
+                      onEnter: (_) => isHovered.value = true,
+                      onExit: (_) => isHovered.value = false,
+                      child: ValueListenableBuilder(
+                          valueListenable: isHovered,
+                          child: Image.file(
+                            File(emojiPack.emojiPath),
                           ),
-                          padding: const EdgeInsets.all(8),
-                          margin: const EdgeInsets.symmetric(horizontal: 16),
-                          child: child,
-                        );
-                      },
+                          builder: (_, isHovered, child) {
+                            return AnimatedContainer(
+                              duration: const Duration(milliseconds: 200),
+                              decoration: BoxDecoration(
+                                color: darkGray100,
+                                borderRadius: BorderRadius.circular(
+                                  isHovered ? 8 : 32,
+                                ),
+                              ),
+                              padding: const EdgeInsets.all(8),
+                              margin: const EdgeInsets.symmetric(horizontal: 16),
+                              child: child,
+                            );
+                          }),
                     ),
                   );
                 },
@@ -80,7 +84,45 @@ class Sidebar extends StatelessWidget {
           ),
           InkWell(
             onTap: () {
-              context.read<EmojiPackBloc>().add(const DeleteAllEmojiPack());
+              showDialog(
+                context: context,
+                builder: (_) {
+                  return AlertDialog(
+                    backgroundColor: darkGray200,
+                    surfaceTintColor: darkGray200,
+                    content: const Text(
+                      'Are you sure you want to delete all emoji packs?',
+                      style: TextStyle(color: Colors.white),
+                    ),
+                    title: const Text(
+                      'Delete All Emoji Packs',
+                      style: TextStyle(
+                        color: Colors.white,
+                        fontSize: 16,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    actions: [
+                      TextButton(
+                        onPressed: () {
+                          Navigator.pop(context);
+                        },
+                        child: const Text('Cancel'),
+                      ),
+                      TextButton(
+                        onPressed: () {
+                          context.read<EmojiPackBloc>().add(const DeleteAllEmojiPack());
+                          Navigator.pop(context);
+                        },
+                        child: const Text(
+                          'Delete All',
+                          style: TextStyle(color: Colors.red),
+                        ),
+                      ),
+                    ],
+                  );
+                },
+              );
             },
             child: Container(
               width: double.infinity,
